requirements.txt
jinja2
PyYAML
rich
transformers